build(flit): add description and version for flit
Flit requires the module file itself hold the description and version
for the package.

I chose the version to correspond to the number of commits listed on
GitHub. This is arbitrary, and as long as the version number increases
and roughly matches semantic versioning[1] (major.minor.patch), it
should be fine to change it to something preferable.

[1]: semver.org